offeron: moved RequestService in the subdirectory
