+++ app/Services/Requests/Exceptions/RequestServiceException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace App\Services\Requests\Exceptions;
+
+class RequestServiceException extends \RuntimeException {
+    //
+}
